Move everything dialer-related in internal/dialer
