+++ internal/dnsconf/dnsconf.go
@@ -2,12 +2,12 @@
 package dnsconf
 
 import (
-	"github.com/ooni/netx/internal/dialerapi"
+	"github.com/ooni/netx/internal/dialer"
 	"github.com/ooni/netx/internal/resolver"
 )
 
 // ConfigureDNS implements netx.Dialer.ConfigureDNS.
-func ConfigureDNS(dialer *dialerapi.Dialer, network, address string) error {
+func ConfigureDNS(dialer *dialer.Dialer, network, address string) error {
 	r, err := resolver.New(dialer.Beginning, dialer.Handler, network, address)
 	if err == nil {
 		dialer.Resolver = r
